+++ packages/build-user-config/src/config/user.config.js
@@ -1,6 +1,6 @@
 const { validation } = require('@builder/app-helpers');
 
-const watchIgnoredRegexp = process.env.RUNTIME_DEBUG ? /node_modules/ : /node_modules|\.ice|\.rax/;
+const watchIgnoredRegexp = process.env.RUNTIME_DEBUG ? /node_modules/ : /node_modules|[/\\]\.ice[/\\]|[/\\]\.rax[/\\]/;
 
 module.exports = [
   {
@@ -193,11 +193,9 @@ module.exports = [
   {
     name: 'eslint',
     validation: (val) => {
-      // for default value
-      if (val === null) return true;
       return validation('eslint', val, 'boolean|object');
     },
-    defaultValue: null,
+    defaultValue: false,
   },
   {
     name: 'tsChecker',