+++ packages/vite-service/src/wp2vite/index.ts
@@ -1,12 +1,13 @@
-import reactRefresh from '@vitejs/plugin-react-refresh';
+import react from '@vitejs/plugin-react';
 import analyzer from 'rollup-plugin-visualizer';
 import * as path from 'path';
 import { all } from 'deepmerge';
 import { isObject } from 'lodash';
 import tsChecker from 'vite-plugin-ts-types';
 import { Context, ITaskConfig } from 'build-scripts';
-import { InlineConfig, BuildOptions } from 'vite';
+import type { TransformOptions } from '@babel/core';
 import eslintReport from 'vite-plugin-eslint-report';
+import type { InlineConfig, BuildOptions, PluginOption } from 'vite';
 import { recordMap } from './config';
 import {
   externalsPlugin,
@@ -16,6 +17,7 @@ import {
   htmlPlugin,
   ignoreHtmlPlugin,
   mockPlugin,
+  ImportDeclarations,
 } from '../plugins';
 
 type Option = BuildOptions & InlineConfig;
@@ -29,6 +31,11 @@ const getAnalyzer = (config: ITaskConfig['chainConfig']) => {
 const getWebpackConfig = (context: Context) => {
   const configArr = context.getWebpackConfig();
   return configArr[0];
+};
+
+// simple array merge for config merge
+const arrayMerge = (destinationArray: any[], sourceArray: any[]) => {
+  return [...(destinationArray || []), ...(sourceArray || [])];
 };
 
 const isBuild = (command: string) => command === 'build';
@@ -98,6 +105,43 @@ const getOpen = (context: Context) => {
   return page ?? true;
 };
 
+type BabelPreset = [string, Record<string, any>];
+
+const getPluginReact = (context: Context): PluginOption[] => {
+  const { userConfig, originalUserConfig, command  } = context;
+  const fastRefresh = userConfig.fastRefresh && command === 'start';
+  const jsxRuntimeConfig = userConfig.babelPresets
+    && (userConfig.babelPresets as BabelPreset[])
+      .find(([pluginName, pluginOptions]) => pluginName.includes('@babel/preset-react') && pluginOptions?.runtime === 'automatic');
+  const babelPlugins = userConfig.babelPlugins as TransformOptions['plugins'];
+  if (fastRefresh || babelPlugins.length > 0 || jsxRuntimeConfig || originalUserConfig.babelPresets) {
+    // get exclude rule for babel-loader
+    return react({
+      // exclude rule for fast refresh
+      exclude: [/node_modules/, /[\\/]\.ice[\\/]/],
+      fastRefresh,
+      jsxRuntime: jsxRuntimeConfig ? 'automatic' : 'classic',
+      babel: {
+        plugins: babelPlugins,
+        // 仅用户配置的 babelPresets 生效
+        // 通过插件修改后的配置如 ['@babel/preset-react', { runtime: 'automatic' }]，通过 jsxRuntime 配置生效
+        presets: originalUserConfig.babelPresets as TransformOptions['presets'],
+      }
+    });
+  }
+  return [];
+};
+
+const getRedirectImport = (context: Context) => {
+  const importDeclarations = context.getValue('importDeclarations') || {};
+  // formate importDeclarations
+  const redirects = Object.keys(importDeclarations).map((name) => ({
+    name,
+    ...importDeclarations[name]
+  })) as ImportDeclarations[];
+  return redirects;
+};
+
 /**
  * Exposed
  */
@@ -113,7 +157,8 @@ export const wp2vite = (context: Context): InlineConfig => {
       getAnalyzer(config.chainConfig),
       // TODO: User Config Type Completion
       externalsPlugin(userConfig.externals as any),
-      importPlugin({ rootDir, entries: userConfig.entry as any }),
+      // import xx from 'ice' 的重定向逻辑
+      importPlugin({ source: 'ice', redirectImports: getRedirectImport(context) }),
       // spa 与 mpa 中对 html 的处理
       serverHistoryPlugin(config.chainConfig.devServer.get('historyApiFallback')),
       getHtmlPlugin(context),
@@ -126,11 +171,12 @@ export const wp2vite = (context: Context): InlineConfig => {
         rootDir,
       }),
       userConfig.ignoreHtmlTemplate ? ignoreHtmlPlugin(rootDir) : null,
+      ...getPluginReact(context),
     ].filter(Boolean),
   };
   if (userConfig.eslint !== false) {
     let eslintOptions = { ignoreInitial: false };
-    if (!userConfig.eslint) {
+    if (userConfig.eslint === true) {
       // lint only changed files, skip lint on start
       eslintOptions.ignoreInitial = true;
     } else {
@@ -148,7 +194,7 @@ export const wp2vite = (context: Context): InlineConfig => {
       recordMap(config.chainConfig, context),
       viteConfig,
       userConfig.vite
-    ]);
+    ], { arrayMerge });
   }
 
   const devServerConfig = {
@@ -184,15 +230,9 @@ export const wp2vite = (context: Context): InlineConfig => {
           include: ['react-app-renderer', 'create-app-shared'],
         },
         server: devServerConfig,
-        plugins: [
-          userConfig.fastRefresh ? reactRefresh({
-            // Exclude node_modules and ice runtime
-            exclude: [/node_modules/, /\.ice/]
-          }) : null
-        ],
       },
       viteConfig,
-    ]);
+    ], { arrayMerge });
   } else {
     return all([{
       build: {
@@ -200,6 +240,6 @@ export const wp2vite = (context: Context): InlineConfig => {
           exclude: ['react-app-renderer', 'create-app-shared'],
         },
       },
-    }, viteConfig]);
+    }, viteConfig], { arrayMerge });
   }
 };