+++ packages/build-user-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.5
+
+- [fix] prevent minimize `.scss` by default.
+- [fix] Regexp for runtime folder
+- [fix] disable default file lint for hmr performance
+
 ## 1.1.4
 
 - [fix] handle unexpected merge with `devServer.allowedHosts`.